+++ config/settings/base.py
@@ -131,7 +131,7 @@ EMAIL_SUBJECT_PREFIX = env('EMAIL_SUBJECT_PREFIX', default='')
 
 # MANAGER CONFIGURATION
 # ------------------------------------------------------------------------------
-ADMINS = [("""Mikko Nieminen""", 'mikko.nieminen@bihealth.de')]
+ADMINS = [("""Admin User""", 'admin.user@example.com')]
 
 # See: https://docs.djangoproject.com/en/1.11/ref/settings/#managers
 MANAGERS = ADMINS
@@ -448,7 +448,7 @@ ENABLED_BACKEND_PLUGINS = env.list(
 # SODAR API settings
 SODAR_API_DEFAULT_VERSION = '0.1'
 SODAR_API_ALLOWED_VERSIONS = [SODAR_API_DEFAULT_VERSION]
-SODAR_API_MEDIA_TYPE = 'application/vnd.bihealth.sodar-core+json'
+SODAR_API_MEDIA_TYPE = 'application/your.application+json'
 SODAR_API_DEFAULT_HOST = env.url(
     'SODAR_API_DEFAULT_HOST', 'http://0.0.0.0:8000'
 )
@@ -476,6 +476,9 @@ PROJECTROLES_DISABLE_CATEGORIES = env.bool(
 )
 PROJECTROLES_INVITE_EXPIRY_DAYS = env.int('PROJECTROLES_INVITE_EXPIRY_DAYS', 14)
 PROJECTROLES_SEND_EMAIL = env.bool('PROJECTROLES_SEND_EMAIL', False)
+PROJECTROLES_EMAIL_SENDER_REPLY = env.bool(
+    'PROJECTROLES_EMAIL_SENDER_REPLY', False
+)
 PROJECTROLES_ENABLE_SEARCH = True
 
 # Optional projectroles settings