+++ docs/source/conf.py
@@ -23,13 +23,13 @@ django.setup()
 # -- Project information -----------------------------------------------------
 
 project = 'SODAR Core'
-copyright = '2019, Berlin Institute of Health'
+copyright = '2020, Berlin Institute of Health'
 author = 'Mikko Nieminen'
 
 # The short X.Y version
 version = '0.7'
 # The full version, including alpha/beta/rc tags
-release = '0.7.1'
+release = '0.7.2'
 
 
 # -- General configuration ---------------------------------------------------