+++ example_backend_app/plugins.py
@@ -25,9 +25,9 @@ class BackendPlugin(BackendPluginPoint):
     #: URL of optional css file to be included
     css_url = 'example_backend_app/css/greeting.css'
 
-    def get_api(self):
+    def get_api(self, **kwargs):
         """Return API entry point object."""
-        return ExampleAPI()
+        return ExampleAPI(**kwargs)
 
     def get_statistics(self):
         return {