+++ projectroles/tests/test_templatetags.py
@@ -495,6 +495,12 @@ class TestProjectrolesTemplateTags(TestTemplateTagsBase):
             0,
         )
 
+    def test_get_project_list_column(self):
+        """Test get_project_list_columns()"""
+        cols = tags.get_project_list_columns()
+        self.assertEqual(len(cols), 2)
+        self.assertEqual(cols[0]['key'], 'links')  # Assert ordering
+
     def test_get_project_column_count(self):
         """Test get_project_column_count()"""
         app_plugins = get_active_plugins()